Fix convergence

+++ vangogh/evolution.py
@@ -133,7 +133,6 @@ class Evolution:
         self.population = selection.select(self.population, self.population_size,
                                            selection_name=self.selection_name)
 
-
     def __umda_generation(self):
         offspring = Population(self.population_size, self.genotype_length, self.initialization)
         offspring.genes[:] = self.population.genes[:]
@@ -142,7 +141,8 @@ class Evolution:
         for i in range(self.genotype_length):
             hist, bins = np.histogram(offspring.genes[:, i], bins=256, range=(0, 256), density=True)
             distribution = hist / np.sum(hist)
-            offspring.genes[:, i] = np.random.choice(np.arange(len(distribution)), size=self.population_size, p=distribution)
+            offspring.genes[:, i] = np.random.choice(np.arange(len(distribution)), size=self.population_size,
+                                                     p=distribution).astype(int)
 
         offspring.fitnesses = drawing_fitness_function(offspring.genes, self.reference_image)
 
@@ -237,4 +237,3 @@ if __name__ == '__main__':
                     noisy_evaluations=False,
                     verbose=True)
     evo.run()
-
